refactor: move strudel-sounds.json from tests/fixtures to src/shared/data
The sound inventory stopped being a test fixture when strudel-validate.ts
started checking every pattern against it. It is now production data that
the (upcoming) out-of-process validation child has to bundle, so it lives
under src/. Still exactly one copy — the guide tests read it from there.
